+++ src/4D/gaugefields_4D_accelerator.jl
@@ -78,7 +78,7 @@ struct Gaugefields_4D_accelerator{NC,TU,TUv,accdevise} <: Gaugefields_4D{NC}
                 @warn "accelerator=\"cuda\" is set but there is no CUDA devise. CPU will be used"
                 U = zeros(ComplexF64, NC, NC, blocksize, rsize)
                 temp_volume = zeros(ComplexF64, blocksize, rsize)
-                accdevise = :threads
+                #accdevise = :threads
                 accdevise = :none
             end
         elseif accelerator == "threads"
@@ -153,11 +153,6 @@ function identityGaugefields_4D_accelerator(NC, NX, NY, NZ, NT, blocks; verbose_
     =#
 end
 
-function set_identity!(U::Gaugefields_4D_accelerator{NC,TU,TUv}) where {NC,TU<:CUDA.CuArray,TUv}
-    CUDA.@sync begin
-        CUDA.@cuda threads = U.blockinfo.blocksize blocks = U.blockinfo.rsize cudakernel_identityGaugefields!(U.U, NC)
-    end
-end
 
 function set_identity!(U::Gaugefields_4D_accelerator{NC,TU,TUv}) where {NC,TU,TUv}
     for r = 1:U.blockinfo.rsize
@@ -172,26 +167,7 @@ function set_identity!(U::Gaugefields_4D_accelerator{NC,TU,TUv}) where {NC,TU,TU
     #end
 end
 
-function cudakernel_randomGaugefields!(U, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_randomGaugefields!(b, r, U.U, NC)
-    #=
-    @inbounds for ic = 1:NC
-        for jc = 1:NC
-            U[jc, ic, b, r] = rand() - 0.5 + im * (rand() - 0.5)
-        end
-    end
-    =#
-end
 
-
-
-function randomize_U!(U::Gaugefields_4D_accelerator{NC,TU,TUv}) where {NC,TU<:CUDA.CuArray,TUv}
-    CUDA.@sync begin
-        CUDA.@cuda threads = U.blockinfo.blocksize blocks = U.blockinfo.rsize cudakernel_randomGaugefields!(U.U, NC)
-    end
-end
 
 function randomize_U!(U::Gaugefields_4D_accelerator{NC,TU,TUv}) where {NC,TU,TUv}
     for r = 1:U.blockinfo.rsize
@@ -228,29 +204,12 @@ function randomGaugefields_4D_accelerator(
     return U
 end
 
-function cudakernel_normalize_U_NC2!(u)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_normalize_U_NC2!(b, r, u)
-    return
-end
 
 
 
-function cudakernel_normalize_U_NC3!(u)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_normalize_U_NC3!(b, r, u)
-    return
-end
 
 
 
-function normalize_U!(U::Gaugefields_4D_accelerator{2,TU,TUv}) where {TU<:CUDA.CuArray,TUv}
-    CUDA.@sync begin
-        CUDA.@cuda threads = U.blockinfo.blocksize blocks = U.blockinfo.rsize cudakernel_normalize_U_NC2!(U.U)
-    end
-end
 
 function normalize_U!(U::Gaugefields_4D_accelerator{2,TU,TUv}) where {TU,TUv}
     for r = 1:U.blockinfo.rsize
@@ -260,11 +219,6 @@ function normalize_U!(U::Gaugefields_4D_accelerator{2,TU,TUv}) where {TU,TUv}
     end
 end
 
-function normalize_U!(U::Gaugefields_4D_accelerator{3,TU,TUv}) where {TU<:CUDA.CuArray,TUv}
-    CUDA.@sync begin
-        CUDA.@cuda threads = U.blockinfo.blocksize blocks = U.blockinfo.rsize cudakernel_normalize_U_NC3!(U.U)
-    end
-end
 
 
 function normalize_U!(U::Gaugefields_4D_accelerator{3,TU,TUv}) where {TU,TUv}
@@ -346,22 +300,8 @@ function shift_U(
 end
 
 
-function cudakernel_mul_NC!(C, A, B, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC!(b, r, C, A, B, NC)
-end
 
 
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::T1,
-    b::T2) where {NC,T1<:Gaugefields_4D_accelerator,T2<:Gaugefields_4D_accelerator,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC!(c.U, a.U, b.U, NC)
-    end
-end
 
 
 function LinearAlgebra.mul!(
@@ -394,17 +334,7 @@ function cudakernel_mul_NC!(C, A, B, α, β, NC)
     return
 end
 
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::T1,
-    b::T2,
-    α::Ta,
-    β::Tb,
-) where {NC,T1<:Gaugefields_4D_accelerator,T2<:Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC!(c.U, a.U, b.U, α, β, NC)
-    end
-end
+
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
@@ -416,7 +346,7 @@ function LinearAlgebra.mul!(
 
     for r = 1:c.blockinfo.rsize
         for b = 1:c.blockinfo.blocksize
-            cudakernel_mul_NC!(b, r, c.U, A.U, B.U, α, β, NC)
+            kernel_mul_NC!(b, r, c.U, A.U, B.U, α, β, NC)
         end
     end
 end
@@ -431,32 +361,16 @@ function LinearAlgebra.mul!(
 
     Threads.@threads for r = 1:c.blockinfo.rsize
         Threads.@threads for b = 1:c.blockinfo.blocksize
-            cudakernel_mul_NC!(b, r, c.U, A.U, B.U, α, β, NC)
+            kernel_mul_NC!(b, r, c.U, A.U, B.U, α, β, NC)
         end
     end
 end
 
 
-function cudakernel_mul_NC3!(C, A, B)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC3!(b, r, C, A, B)
-    return
-end
 
 
 
 
-
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{3,TU,TUv},
-    a::T1,
-    b::T2,
-) where {T1<:Gaugefields_4D_accelerator,T2<:Gaugefields_4D_accelerator,TU<:CUDA.CuArray,TUv}
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC3!(c.U, a.U, b.U)
-    end
-end
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{3,TU,TUv,:none},
@@ -482,21 +396,9 @@ function LinearAlgebra.mul!(
     end
 end
 
-function cudakernel_mul_NC_abdag!(C, A, B, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_abdag!(b, r, C, A, B, NC)
-end
 
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    A::T1,
-    B::Adjoint_Gaugefields{T2}) where {NC,T1<:Gaugefields_4D_accelerator,T2<:Gaugefields_4D_accelerator,TU<:CUDA.CuArray,TUv}
 
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_abdag!(c.U, A.U, B.parent.U, NC)
-    end
-end
+
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:threads},
@@ -523,25 +425,9 @@ function LinearAlgebra.mul!(
 end
 
 
-function cudakernel_mul_NC3!(C, A, B, α, β)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC3!(b, r, C, A, B, α, β)
-    return
-end
 
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{3,TU,TUv},
-    a::T1,
-    b::T2,
-    α::Ta,
-    β::Tb,
-) where {T1<:Gaugefields_4D_accelerator,T2<:Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
 
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC3!(c.U, a.U, b.U, α, β, NC)
-    end
-end
+
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{3,TU,TUv,:none},
@@ -571,23 +457,7 @@ function LinearAlgebra.mul!(
     end
 end
 
-function cudakernel_mul_NC_abdag!(C, A, B, α, β, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_abdag!(b, r, C, A, B, α, β, NC)
-end
 
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::T1,
-    b::Adjoint_Gaugefields{T2},
-    α::Ta,
-    β::Tb
-) where {NC,T1<:Gaugefields_4D_accelerator,T2<:Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_abdag!(c.U, a.U, b.parent.U, α, β, NC)
-    end
-end
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
@@ -617,25 +487,7 @@ function LinearAlgebra.mul!(
     end
 end
 
-function cudakernel_mul_NC_adagbdag!(C, A, B, α, β, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_adagbdag!(b, r, C, A, B, α, β, NC)
-end
 
-
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::Adjoint_Gaugefields{T1},
-    b::Adjoint_Gaugefields{T2},
-    α::Ta,
-    β::Tb
-) where {NC,T1<:Gaugefields_4D_accelerator,T2<:Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_adagbdag!(c.U, a.parent.U, b.parent.U, α, β, NC)
-    end
-end
 
 #c = A'* B'
 function LinearAlgebra.mul!(
@@ -669,24 +521,6 @@ function LinearAlgebra.mul!(
 end
 
 
-function cudakernel_mul_NC_adagbdag!(C, A, B, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_adagbdag!(b, r, C, A, B, NC)
-end
-
-
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::Adjoint_Gaugefields{T1},
-    b::Adjoint_Gaugefields{T2}
-) where {NC,T1<:Gaugefields_4D_accelerator,T2<:Gaugefields_4D_accelerator,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_adagbdag!(c.U, a.parent.U, b.parent.U, NC)
-    end
-
-end
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
@@ -716,26 +550,6 @@ function LinearAlgebra.mul!(
 
 end
 
-function cudakernel_mul_NC_adagb!(C, A, B, α, β, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_adagb!(b, r, C, A, B, α, β, NC)
-end
-
-
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::Adjoint_Gaugefields{T1},
-    b::T2,
-    α::Ta,
-    β::Tb
-) where {NC,T1<:Gaugefields_4D_accelerator,T2<:Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_adagb!(c.U, a.parent.U, b.U, α, β, NC)
-    end
-
-end
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
@@ -773,27 +587,6 @@ end
 
 
 
-function cudakernel_mul_NC_abshift!(C, A, B, α, β, shift, blockinfo::Blockindices, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_abshift!(b, r, C, A, B, α, β, shift, blockinfo, NC)
-end
-
-
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::T1,
-    b::T2,
-    α::Ta,
-    β::Tb
-) where {NC,T1<:Gaugefields_4D_accelerator,T2<:Shifted_Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_abshift!(c.U, a.U, b.parent.U, α, β,
-            b.shift, b.parent.blockinfo, NC)
-    end
-end
-
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
     A::T1,
@@ -825,28 +618,6 @@ function LinearAlgebra.mul!(
 end
 
 
-function cudakernel_mul_NC_ashiftb!(C, A, B, α, β, shift, blockinfo::Blockindices, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_ashiftb!(b, r, C, A, B, α, β, shift, blockinfo, NC)
-end
-
-
-
-
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::T1,
-    b::T2,
-    α::Ta,
-    β::Tb
-) where {NC,T1<:Shifted_Gaugefields_4D_accelerator,T2<:Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_ashiftb!(c.U, a.parent.U, b.U, α, β,
-            a.shift, a.parent.blockinfo, NC)
-    end
-end
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
@@ -882,26 +653,6 @@ function LinearAlgebra.mul!(
 
 end
 
-function cudakernel_mul_NC_ashiftbshift!(C, A, B, α, β, ashift, bshift, blockinfo::Blockindices, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_ashiftbshift!(b, r, C, A, B, α, β, ashift, bshift, blockinfo, NC)
-end
-
-
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::T1,
-    b::T2,
-    α::Ta,
-    β::Tb
-) where {NC,T1<:Shifted_Gaugefields_4D_accelerator,T2<:Shifted_Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_ashiftbshift!(c.U, a.parent.U, b.parent.U, α, β,
-            a.shift, b.shift, a.parent.blockinfo, NC)
-    end
-end
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
@@ -935,30 +686,6 @@ function LinearAlgebra.mul!(
         end
     end
 
-end
-
-function cudakernel_mul_NC_ashiftbshiftdag!(C, A, B, α, β, ashift, bshift, blockinfo::Blockindices, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_ashiftbshiftdag!(b, r, C, A, B, α, β, ashift, bshift, blockinfo, NC)
-end
-
-
-
-
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::T1,
-    b::Adjoint_Gaugefields{T2},
-    α::Ta,
-    β::Tb
-) where {NC,T1<:Shifted_Gaugefields_4D_accelerator,T2<:Shifted_Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_ashiftbshiftdag!(
-            c.U, a.parent.U, b.parent.parent.U, α, β,
-            a.shift, b.parent.shift, a.parent.blockinfo, NC)
-    end
 end
 
 function LinearAlgebra.mul!(
@@ -1000,27 +727,6 @@ end
 
 
 
-function cudakernel_mul_NC_adagbshift!(C, A, B, α, β, shift, blockinfo::Blockindices, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_adagbshift!(b, r, C, A, B, α, β, shift, blockinfo, NC)
-end
-
-
-
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::Adjoint_Gaugefields{T1},
-    b::T2,
-    α::Ta,
-    β::Tb
-) where {NC,T1<:Gaugefields_4D_accelerator,T2<:Shifted_Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_adagbshift!(c.U, a.parent.U, b.parent.U, α, β,
-            b.shift, b.parent.blockinfo, NC)
-    end
-end
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
@@ -1054,26 +760,6 @@ function LinearAlgebra.mul!(
     end
 end
 
-function cudakernel_mul_NC_adagbshiftdag!(C, A, B, α, β, shift, blockinfo::Blockindices, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_adagbshiftdag!(b, r, C, A, B, α, β, shift, blockinfo, NC)
-end
-
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::Adjoint_Gaugefields{T1},
-    b::Adjoint_Gaugefields{T2},
-    α::Ta,
-    β::Tb
-) where {NC,T1<:Gaugefields_4D_accelerator,T2<:Shifted_Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_adagbshiftdag!(
-            c.U, a.parent.U, b.parent.parent.U, α, β,
-            b.parent.shift, b.parent.parent.blockinfo, NC)
-    end
-end
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
@@ -1111,27 +797,6 @@ function LinearAlgebra.mul!(
 
 end
 
-function cudakernel_mul_NC_ashiftbdag!(C, A, B, α, β, shift, blockinfo::Blockindices, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_ashiftbdag!(b, r, C, A, B, α, β, shift, blockinfo, NC)
-end
-
-
-
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::T1,
-    b::Adjoint_Gaugefields{T2},
-    α::Ta,
-    β::Tb
-) where {NC,T1<:Shifted_Gaugefields_4D_accelerator,T2<:Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_ashiftbdag!(c.U, a.parent.U, b.parent.U, α, β,
-            a.shift, a.parent.blockinfo, NC)
-    end
-end
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
@@ -1166,27 +831,7 @@ function LinearAlgebra.mul!(
 end
 
 
-function cudakernel_mul_NC_ashiftdagbdag!(C, A, B, α, β, shift, blockinfo::Blockindices, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_ashiftdagbdag!(b, r, C, A, B, α, β, shift, blockinfo, NC)
-end
 
-
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::Adjoint_Gaugefields{T1},
-    b::Adjoint_Gaugefields{T2},
-    α::Ta,
-    β::Tb
-) where {NC,T1<:Shifted_Gaugefields_4D_accelerator,T2<:Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_ashiftdagbdag!(
-            c.U, a.parent.parent.U, b.parent.U, α, β,
-            a.parent.shift, a.parent.parent.blockinfo, NC)
-    end
-end
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
@@ -1224,27 +869,7 @@ function LinearAlgebra.mul!(
 
 end
 
-function cudakernel_mul_NC_abshiftdag!(C, A, B, α, β, shift, blockinfo::Blockindices, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_abshiftdag!(b, r, C, A, B, α, β, shift, blockinfo, NC)
-end
 
-
-
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::T1,
-    b::Adjoint_Gaugefields{T2},
-    α::Ta,
-    β::Tb
-) where {NC,T1<:Gaugefields_4D_accelerator,T2<:Shifted_Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_abshiftdag!(c.U, a.U, b.parent.parent.U, α, β,
-            b.parent.shift, b.parent.parent.blockinfo, NC)
-    end
-end
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
@@ -1279,28 +904,6 @@ function LinearAlgebra.mul!(
 end
 
 
-function cudakernel_mul_NC_ashiftdagb!(C, A, B, α, β, shift, blockinfo::Blockindices, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_ashiftdagb!(b, r, C, A, B, α, β, shift, blockinfo, NC)
-end
-
-
-
-
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::Adjoint_Gaugefields{T1},
-    b::T2,
-    α::Ta,
-    β::Tb
-) where {NC,T1<:Shifted_Gaugefields_4D_accelerator,T2<:Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_ashiftdagb!(c.U, a.parent.parent.U, b.U, α, β,
-            a.parent.shift, a.parent.parent.blockinfo, NC)
-    end
-end
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
@@ -1337,26 +940,7 @@ function LinearAlgebra.mul!(
 end
 
 
-function cudakernel_mul_NC_ashiftdagbshiftdag!(C, A, B, α, β, ashift, bshift, blockinfo::Blockindices, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_ashiftdagbshiftdag!(b, r, C, A, B, α, β, ashift, bshift, blockinfo, NC)
-end
 
-
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::Adjoint_Gaugefields{T1},
-    b::Adjoint_Gaugefields{T2},
-    α::Ta,
-    β::Tb
-) where {NC,T1<:Shifted_Gaugefields_4D_accelerator,T2<:Shifted_Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_ashiftdagbshiftdag!(c.U, a.parent.parent.U, b.parent.parent.U, α, β,
-            a.parent.shift, b.parent.shift, a.parent.parent.blockinfo, NC)
-    end
-end
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
@@ -1390,25 +974,7 @@ function LinearAlgebra.mul!(
     end
 end
 
-function cudakernel_mul_NC_ashiftdagbshift!(C, A, B, α, β, ashift, bshift, blockinfo::Blockindices, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_mul_NC_ashiftdagbshift!(b, r, C, A, B, α, β, ashift, bshift, blockinfo, NC)
-end
 
-function LinearAlgebra.mul!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    a::Adjoint_Gaugefields{T1},
-    b::T2,
-    α::Ta,
-    β::Tb
-) where {NC,T1<:Shifted_Gaugefields_4D_accelerator,T2<:Shifted_Gaugefields_4D_accelerator,Ta<:Number,Tb<:Number,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_mul_NC_ashiftdagbshift!(c.U, a.parent.parent.U, b.parent.U, α, β,
-            a.parent.shift, b.shift, a.parent.parent.blockinfo, NC)
-    end
-end
 
 function LinearAlgebra.mul!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
@@ -1445,24 +1011,7 @@ end
 
 
 
-function cudakernel_tr!(temp_volume, U, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_tr!(b, r, temp_volume, U, NC)
-    return
-end
 
-
-
-function LinearAlgebra.tr(a::Gaugefields_4D_accelerator{NC,TU,TUv}) where {NC,TU<:CUDA.CuArray,TUv}
-    CUDA.@sync begin
-        CUDA.@cuda threads = a.blockinfo.blocksize blocks = a.blockinfo.rsize cudakernel_tr!(a.temp_volume, a.U, NC)
-    end
-
-    s = CUDA.reduce(+, a.temp_volume)
-    return s
-
-end
 
 function LinearAlgebra.tr(a::Gaugefields_4D_accelerator{NC,TU,TUv,:none}) where {NC,TU,TUv}
     for r = 1:a.blockinfo.rsize
@@ -1486,27 +1035,7 @@ function LinearAlgebra.tr(a::Gaugefields_4D_accelerator{NC,TU,TUv,:threads}) whe
 
 end
 
-function cudakernel_tr!(temp_volume, A, B, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_tr!(b, r, temp_volume, A, B, NC)
-    return
-end
 
-
-function LinearAlgebra.tr(
-    a::Gaugefields_4D_accelerator{NC,TU,TUv},
-    b::Gaugefields_4D_accelerator{NC,TU,TUv},
-) where {NC,TU<:CUDA.CuArray,TUv}
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = a.blockinfo.blocksize blocks = a.blockinfo.rsize cudakernel_tr!(a.temp_volume, a.U, b.U, NC)
-    end
-
-    s = CUDA.reduce(+, a.temp_volume)
-
-    return s
-end
 
 function LinearAlgebra.tr(
     A::Gaugefields_4D_accelerator{NC,TU,TUv,:none},
@@ -1542,25 +1071,7 @@ end
 
 
 
-function substitute_U!(A::Gaugefields_4D_accelerator{NC,TU,TUv}, B::Gaugefields_4D_nowing{NC}) where {NC,TU<:CUDA.CuArray,TUv}
-    acpu = Array(A.U)
 
-    blockinfo = A.blockinfo
-    for r = 1:blockinfo.rsize
-        for b = 1:blockinfo.blocksize
-            ix, iy, iz, it = fourdim_cordinate(b, r, blockinfo)
-            #println((ix,iy,iz,it))
-            for ic = 1:NC
-                for jc = 1:NC
-                    acpu[jc, ic, b, r] = B[jc, ic, ix, iy, iz, it]
-                end
-            end
-        end
-    end
-    agpu = CUDA.CuArray(acpu)
-    A.U .= agpu
-
-end
 
 function substitute_U!(A::Gaugefields_4D_accelerator{NC,TU,TUv}, B::Gaugefields_4D_nowing{NC}) where {NC,TU,TUv}
     acpu = Array(A.U)
@@ -1618,37 +1129,9 @@ function substitute_U!(A::Gaugefields_4D_nowing{NC}, B::Gaugefields_4D_accelerat
     end
 end
 
-function substitute_U!(A::Gaugefields_4D_nowing{NC}, B::Gaugefields_4D_accelerator{NC,TU,TUv}) where {NC,TU<:CUDA.CuArray,TUv}
-    bcpu = Array(B.U)
-
-    blockinfo = B.blockinfo
-    for r = 1:blockinfo.rsize
-        for b = 1:blockinfo.blocksize
-            ix, iy, iz, it = fourdim_cordinate(b, r, blockinfo)
-
-            for ic = 1:NC
-                for jc = 1:NC
-                    A[jc, ic, ix, iy, iz, it] = bcpu[jc, ic, b, r]
-                end
-            end
-        end
-    end
-end
-
-function cudakernel_add_U!(c, a, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_add_U!(b, r, c, a, NC)
-    return
-end
 
 
 
-function add_U!(c::Gaugefields_4D_accelerator{NC,TU,TUv}, a::T1) where {NC,T1<:Gaugefields_4D_accelerator,TU<:CUDA.CuArray,TUv}
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_add_U!(c.U, a.U, NC)
-    end
-end
 
 function add_U!(c::Gaugefields_4D_accelerator{NC,TU,TUv}, a::T1) where {NC,T1<:Gaugefields_4D_accelerator,TU,TUv}
     for r = 1:c.blockinfo.rsize
@@ -1658,24 +1141,7 @@ function add_U!(c::Gaugefields_4D_accelerator{NC,TU,TUv}, a::T1) where {NC,T1<:G
     end
 end
 
-function cudakernel_add_U_αa!(c, a, α, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_add_U_αa!(b, r, c, a, α, NC)
-    return
-end
 
-
-
-function add_U!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    α::N,
-    a::T1,
-) where {NC,T1<:Gaugefields_4D_accelerator{NC},N<:Number,TU<:CUDA.CuArray,TUv}
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_add_U_αa!(c.U, a.U, α, NC)
-    end
-end
 
 function add_U!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv},
@@ -1689,27 +1155,7 @@ function add_U!(
     end
 end
 
-function cudakernel_add_U_αshifta!(c, a, α, shift, blockinfo, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_add_U_αshifta!(b, r, c, a, α, shift, blockinfo, NC)
-    return
-end
 
-
-
-
-function add_U!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    α::N,
-    a::T1,
-) where {NC,T1<:Shifted_Gaugefields_4D_accelerator,N<:Number,TU<:CUDA.CuArray,TUv}
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_add_U_αshifta!(
-            c.U, a.parent.U, α,
-            a.shift, a.parent.blockinfo, NC)
-    end
-end
 
 function add_U!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv},
@@ -1725,25 +1171,7 @@ function add_U!(
     end
 end
 
-function cudakernel_add_U_αadag!(c, a, α, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_add_U_αadag!(b, r, c, a, α, NC)
-    return
-end
 
-
-
-
-function add_U!(
-    c::Gaugefields_4D_accelerator{NC,TU,TUv},
-    α::N,
-    a::Adjoint_Gaugefields{T1},
-) where {NC,T1<:Gaugefields_4D_accelerator{NC},N<:Number,TU<:CUDA.CuArray,TUv}
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_add_U_αadag!(c.U, a.U, α, NC)
-    end
-end
 
 function add_U!(
     c::Gaugefields_4D_accelerator{NC,TU,TUv},
@@ -1758,20 +1186,6 @@ function add_U!(
     end
 end
 
-function cudakernel_clear_U!(c, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_clear_U!(b, r, c, NC)
-    return
-end
-
-
-
-function clear_U!(c::Gaugefields_4D_accelerator{NC,TU,TUv}) where {NC,TU<:CUDA.CuArray,TUv}
-    CUDA.@sync begin
-        CUDA.@cuda threads = c.blockinfo.blocksize blocks = c.blockinfo.rsize cudakernel_clear_U!(c.U, NC)
-    end
-end
 
 function clear_U!(c::Gaugefields_4D_accelerator{NC,TU,TUv}) where {NC,TU,TUv}
     for r = 1:c.blockinfo.rsize
@@ -1792,13 +1206,6 @@ function LinearAlgebra.mul!(
     LinearAlgebra.mul!(c, a, b, 1, 0)
 end
 
-function cudakernel_exptU_wvww!(w, v, ww, t, NC)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_exptU_wvww!(b, r, w, v, ww, t, NC)
-    return
-
-end
 
 
 function exptU!(
@@ -1819,22 +1226,7 @@ function exptU!(
     error("exptU with general NC is not implemented")
 end
 
-function exptU!(
-    uout::T,
-    t::N,
-    v::Gaugefields_4D_accelerator{3,TU,TUv},
-    temps::Array{T,1},
-) where {N<:Number,T<:Gaugefields_4D_accelerator,TU<:CUDA.CuArray,TUv} #uout = exp(t*u)
 
-    ww = temps[1]
-    w = temps[2]
-
-    CUDA.@sync begin
-        CUDA.@cuda threads = v.blockinfo.blocksize blocks = v.blockinfo.rsize cudakernel_exptU_wvww!(w.U, v.U, ww.U, t, NC)
-    end
-
-    mul!(uout, w', ww)
-end
 
 
 function exptU!(
@@ -1871,34 +1263,7 @@ function mul_skiplastindex!(
 
 end
 
-function cudakernel_Traceless_antihermitian_NC3!(vout, vin)
-    b = Int64(CUDA.threadIdx().x)
-    r = Int64(CUDA.blockIdx().x)
-    kernel_Traceless_antihermitian_NC3!(b, r, vout, vin)
-    return
-end
 
-"""
------------------------------------------------------c
-     !!!!!   vin and vout should be different vectors
-
-     Projectin of the etraceless antiermite part 
-     vout = x/2 - Tr(x)/6
-     wher   x = vin - Conjg(vin)      
------------------------------------------------------c
-    """
-
-#Q = -(1/2)*(Ω' - Ω) + (1/(2NC))*tr(Ω' - Ω)*I0_2
-#Omega' - Omega = -2i imag(Omega)
-function Traceless_antihermitian!(
-    vout::Gaugefields_4D_accelerator{3,TU,TUv},
-    vin::Gaugefields_4D_accelerator{3},
-) where {TU<:CUDA.CuArray,TUv}
-    CUDA.@sync begin
-        CUDA.@cuda threads = vout.blockinfo.blocksize blocks = vout.blockinfo.rsize cudakernel_Traceless_antihermitian_NC3!(vout.U, vin.U)
-    end
-
-end
 
 #Q = -(1/2)*(Ω' - Ω) + (1/(2NC))*tr(Ω' - Ω)*I0_2
 #Omega' - Omega = -2i imag(Omega)
